# code/web/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "App API",
        "version": "0.1"
    },
    "paths": {
...
        "/orders/{id}": {
            "get": {
                "tags": [
                    "orders"
                ],
                "operationId": "da55625fa2c37a9d466686f574d9b4cd",
                "parameters": [
                    {
                        "parameter": "id",
                        "name": "id",
                        "in": "path",
                        "description": "order id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "items": {}
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/orders/append": {
            "post": {
                "tags": [
                    "orders"
                ],
                "operationId": "7925b4d30406731ef849658082e89fde",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_id",
                                    "dish_id"
                                ],
                                "properties": {
                                    "order_id": {
                                        "type": "integer"
                                    },
                                    "dish_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "dish appended",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}